Avoid throwing away decoded but unplayed buffers at the end of a track.
Change tracks on end_of_track callback instead of trying to calculate the
current position; this avoids forcing a track change too early. Also avoid
throwing away data in the buffers when we automatically change to the next
track since we probably want to listen to the tracks to the end.